rebase: move code to new yurtmanager location
